backend/pyproject.toml
[tool.poetry]
name = "satsverdant-backend"
version = "1.0.0"
description = "SatsVerdant waste verification backend API"
authors = ["SatsVerdant Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
redis = "^5.0.1"
celery = "^5.3.4"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
httpx = "^0.25.2"
google-cloud-storage = "^2.10.0"
python-dotenv = "^1.0.0"
opencv-python-headless = "^4.8.1"
scikit-learn = "^1.3.2"
numpy = "^1.24.3"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"
pillow = "^10.1.0"
imagehash = "^4.3.1"
requests = "^2.31.0"

[tool.poetry.group.ml]
optional = true

[tool.poetry.group.ml.dependencies]
# TensorFlow for model training (optional - not needed for API server)
# Install with: poetry install --with ml
tensorflow = "^2.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
faker = "^20.1.0"
factory-boy = "^3.3.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | __pycache__
  | alembic
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=app --cov-report=term-missing --cov-report=html"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "alembic",
]

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/alembic/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
